% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarietyDevPipe.R
\name{addPhenoRecords}
\alias{addPhenoRecords}
\title{addPhenoRecords function}
\usage{
addPhenoRecords(pheno, trialType, bsd)
}
\arguments{
\item{pheno}{Named vector of phenotypic values}

\item{trialType}{String name of the trial type to get the error variance}

\item{bsd}{The breeding scheme data list}
}
\value{
Updates bsd with new phenotypes in phenoRecords
}
\description{
Function to add new phenotypes from a VDP trial to phenoRecords
}
\details{
Call this function from the runVDPstage
}
\examples{
newPhenoRec <- addPhenoRecords(newPheno, "CET", bsd)

}
