% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprov.R
\name{calcGRM}
\alias{calcGRM}
\title{calcGRM function}
\usage{
calcGRM(bsd)
}
\arguments{
\item{bsd}{List of breeding scheme data}

\item{SP}{The AlphaSimR SimParam object. Needed to pull the SNP genotypes}
}
\value{
A genomic relationship matrix
}
\description{
Function to make a genomic relationship matrix to calculat GEBVs
}
\details{
bsd contains both variety candidates that have phenotypes and
breeding population individuals that don't.  Both need to be in the GRM
for prediction
}
\examples{
grm <- calcGRM(bsd)

}
