% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprov.R
\name{convertNamesToRows}
\alias{convertNamesToRows}
\title{convertNamesToRows function}
\usage{
convertNamesToRows(nameMat)
}
\arguments{
\item{nameMat}{The pedigree matrix with character IDs}
}
\value{
Matrix suitable for pedigreeToCCmatrix
}
\description{
Function to transform arbitrary name IDs to row numbers for the sake of
pedigreeToCCmatrix
}
\details{
Keeps everything in order
}
\examples{
threeColPed <- convertNamesToRows(nameMat)]

}
