% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprov.R
\name{grmPhenoEval}
\alias{grmPhenoEval}
\title{grmPhenoEval function}
\usage{
grmPhenoEval(bsd)
}
\arguments{
\item{bsd}{List of breeding scheme data}
}
\value{
Named real vector of the GEBVs (names are the individual ids) of all
individuals including breeding population and variety candidates,
with appropriate weights by error variance of the observations
}
\description{
Takes phenotypes from phenoRecords and combines them with the GRM across
variety candidates and breeding population individuals
}
\details{
Given all the phenotypic records calculate the GEBV
}
\examples{
grmBLUPs <- grmPhenoEval(bsd)

}
