% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarietyDevPipe.R
\name{iidPhenoEval}
\alias{iidPhenoEval}
\title{iidPhenoEval function}
\usage{
iidPhenoEval(phenoRecords)
}
\arguments{
\item{phenoRecords}{Tibble of phenotypic observations on variety candidates.}
}
\value{
Named real vector of the BLUPs of all individuals in phenoRecords,
(names are the individual ids), with appropriate weights by error variance
}
\description{
Function to estimate variety candidate BLUPs assuming they are IID
}
\details{
Given all the phenotypic records calculate the best prediction of
the genotypic value for each individual using all its records
}
\examples{
iidBLUPs <- iidPhenoEval(phenoRecords)

}
