% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{initializeProgram}
\alias{initializeProgram}
\title{initializeProgram function}
\usage{
initializeProgram(founderFile, schemeFile)
}
\arguments{
\item{founderFile}{String name of the text file with parameters governing
the founders and the genetic architecture}

\item{schemeFile}{String name of the text file with parameters governing
the variety development pipeline and selection functions}
}
\value{
A named list containing the `AlphaSimR` SP file, bsd a list of
breeding scheme data, breedingPop the breeding population, varietyCand
an empty pop for variety candidates, phenoRecords an empty tibble for
phenotype records, and inventory an empty tibble for variety candidates
}
\description{
Function to initialize the simulation
Read the founder genetic architecture and population genetic parameters
Create simulation data structures
}
\details{
Call this function at the beginning of the simulation
}
\examples{
bsd <- initializeProgram("FounderCtrlFile.txt")

}
