% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprov.R
\name{pedigreeToCCmatrix}
\alias{pedigreeToCCmatrix}
\title{pedigreeToCCmatrix function}
\usage{
pedigreeToCCmatrix(threeColPed)
}
\arguments{
\item{threeColPed}{The pedigree matrix}
}
\value{
Matrix with coefficients of coancestry
}
\description{
Function to calculate a coefficient of coancestry matrix from a three-
column pedigree matrix:
The first column has to be the row number
Sire and dam columns refer directly to rows
Unknown parents need to be set to 0 (zero)
}
\details{
Uses the breedingPop data in bsd
}
\examples{
ccMat <- pedigreeToCCmatrix(threeColPed)]

}
