% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarietyDevPipe.R
\name{runVDPtrial}
\alias{runVDPtrial}
\title{runVDPtrial function}
\usage{
runVDPtrial(trialType, entries, bsd)
}
\arguments{
\item{trialType}{String name of the kind of trial to run. Must link back to
a trial type with parameters in bsd}

\item{entries}{A vector of entry IDs}

\item{bsd}{The breeding scheme data list}
}
\value{
Updated bsd
}
\description{
Function to run a variety development pipeline stage
}
\details{
Add records to phenoRecords and update the inventory
}
\examples{
params <- runVDPtrial("CET", c("1", "2", "3", "4"), bsd)

}
