% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprov.R
\name{selectParents}
\alias{selectParents}
\title{selectParents function}
\usage{
selectParents(nToSelect, bsd)
}
\arguments{
\item{nToSelect}{The number of IDs to return that have the highest BLUPs}

\item{bsd}{The breeding scheme data list}
}
\value{
List with entries vector of variety IDs selected and updated bsd
}
\description{
Function to select which breeding population individuals to use as parents
to advance the population
}
\details{
Accesses all data in phenoRecords to pick the highest among
candidates with enough seed inventory.
}
\examples{
optCont <- selectParents(nToSelect, bsd)]

}
