% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationImprov.R
\name{selectParents}
\alias{selectParents}
\title{selectParents function}
\usage{
selectParents(nToSelect, bsd)
}
\arguments{
\item{nToSelect}{The number of IDs to return that have the highest BLUPs}

\item{bsd}{The breeding scheme data list}
}
\value{
Data.frame with pedigree ccMat-based optimal contributions
}
\description{
Function to select which breeding population individuals to use as parents
to advance the population
}
\details{
Accesses phenoRecords to predict GEBVs among breeding population
then calculates optimal contributions with those GEBVs and a pedigree-based
coefficient of coancestry matrix.
}
\examples{
optCont <- selectParents(nToSelect, bsd)]

}
